pyproject.toml
[project]
name = "beku-stackabletech"
version = "0.0.6-dev"
authors = [
  { name="Razvan Mihai", email="razvan.mihai@stackable.tech" }
]
description = "Test suite expander for Stackable Kuttl tests."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Jinja2>=3.1.2",
    "PyYAML>=6.0"
]
[project.optional-dependencies]
lint = ['pre-commit', 'pylint', 'mypy']

[project.scripts]
beku = "beku:main"

[project.urls]
"Homepage" = "https://github.com/stackabletech/beku.py"
"Bug Tracker" = "https://github.com/stackabletech/beku.py/issues"
